+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## UNRELEASED
+
+ - Correctly mark duration parameter for status check action as required.
+
 ## v1.0.0
 
  - Initial release